+++ FableCricket/Cricket/View.fs
@@ -1,10 +1,7 @@
 module Cricket.View
 
-open Fable.Core
 open Fable.Helpers.React
 open Fable.Helpers.React.Props
-
-open Elmish
 
 open Cricket.CricketEngine
 open Cricket.CricketEngine.Formatting